feat(stash): single-page and single-session stashes open straight to the content
The single-file-upload shortcut already opens a one-file stash directly on
the file preview (skipping bundle chrome + empty viz). The same intuition
applies to a stash that is just one page (`stash publish notes.md`) or
one session (`stash share <session>`) — there's no bundle to chrome,
the artifact IS the stash.

- primaryFileItemForStash → primaryItemForStash returning a discriminated
  union { kind: "file" | "page" | "session", item }. File rules unchanged;
  page and session require the stash to contain exactly that one item.
- StashPageClient dispatches to SingleFilePreview / PageBody / SessionBody.
- PageBody and SessionBody extracted from StashItemClient.tsx into a
  shared StashItemBodies.tsx so the inline-preview path and the
  /items/{type}/{id} viewer render the same content the same way.
